hopefully fix setting seasons list view height
github issue #30. some people were reporting that the seasons list view
is now not shown. I can't reproduce but I'm guessing that the cause is
the way that we try to make the listview tall enough to contain all its
children. so try a different approach to setting its height.